Revert to Debian "bullseye"
This fixes a regression we found in cURL v7.85.0 and above where it auto-resolves any "*.localhost" domains to the PHP container, not the Apache container as set in /etc/hosts. We couldn't find a way to alter or override this new behaviour, so we're reverting to an older version of Debian (and cURL) as a workaround.